requirements.txt
asgiref==3.8.1
Django==5.1.3
django-tinymce==4.1.0
pillow==11.0.0
sqlparse==0.5.2
